+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Filename sanitization is now optional (enabled via default)
 - Address parsing improved and extended to include more cases
 - Boundary parsing fixed and improved to support more formats #544
+- Decode partially encoded address names #511
 
 ### Added
 - Security configuration options added